+++ src/ga_session.hpp
@@ -101,6 +101,7 @@ namespace sdk {
         nlohmann::json cancel_twofactor_reset(const nlohmann::json& twofactor_data);
 
         nlohmann::json encrypt_with_pin(const nlohmann::json& details);
+        nlohmann::json decrypt_with_pin(const nlohmann::json& details);
         void disable_all_pin_logins();
 
         nlohmann::json get_unspent_outputs(const nlohmann::json& details, unique_pubkeys_and_scripts_t& missing);
@@ -246,6 +247,7 @@ namespace sdk {
         void subscribe_all(locker_t& locker);
 
         std::vector<unsigned char> get_pin_password(const std::string& pin, const std::string& pin_identifier);
+        nlohmann::json decrypt_with_pin_impl(const nlohmann::json& details, bool is_login);
 
         // Start/stop background header downloads
         void download_headers_ctl(locker_t& locker, bool do_start);