+++ src/session_impl.hpp
@@ -163,6 +163,7 @@ namespace sdk {
         virtual nlohmann::json cancel_twofactor_reset(const nlohmann::json& twofactor_data) = 0;
 
         virtual nlohmann::json encrypt_with_pin(const nlohmann::json& details) = 0;
+        virtual nlohmann::json decrypt_with_pin(const nlohmann::json& details);
 
         virtual bool encache_blinding_data(const std::string& pubkey_hex, const std::string& script_hex,
             const std::string& nonce_hex, const std::string& blinding_pubkey_hex);