+++ src/swift/GreenAddress/Sources/GreenAddress/GreenAddress.swift
@@ -360,6 +360,10 @@ public class Session {
         return try jsonFuncToCallHandlerWrapper(input: details, fun: GA_encrypt_with_pin)
     }
 
+    public func decryptWithPin(details: [String: Any]) throws -> TwoFactorCall {
+        return try jsonFuncToCallHandlerWrapper(input: details, fun: GA_decrypt_with_pin)
+    }
+
     public func disableAllPinLogins() throws -> Void {
         try callWrapper(fun: GA_disable_all_pin_logins(session))
     }